eyetracking/requirements.txt
Django==1.8.5
django-debug-toolbar==1.3.2
Pillow==3.0.0
django-jfu==2.0.9
django-uuid-upload-path==1.0.0
django-resized==0.3.5


# Testing
autopep8==1.2.1
flake8==2.4.1
coverage==4.0
